requirements.txt
flask
gunicorn

blinker==1.9.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.3
fonttools==4.62.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.3.5
packaging==26.0
pandas==2.3.3
pillow==12.1.1
psycopg2-binary==2.9.11
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.6
wheel==0.45.1
flask-cors
flask-jwt-extended
